feat(Auth): Rename `resolveIdToken` to `resolveToken`
As it's not necessary an _id token_ that's returned
