+++ auth/resolve-token.js
